[#100] refactor : 모달 컴포넌트 이름 변경
