--- app/(settingsMenu)/logout/page.tsx
@@ -1,7 +1,0 @@
-import { redirect } from "next/navigation";
-
-const Page = () => {
-  redirect("/settings");
-};
-
-export default Page;
