--- app/settings/@modal/(..)logout/page.tsx
@@ -1,12 +1,0 @@
-import Logout from "@/components/Setting/Logout";
-import Modal from "@/components/@common/Modal";
-
-const Page = () => {
-  return (
-    <Modal>
-      <Logout />
-    </Modal>
-  );
-};
-
-export default Page;
