--- app/settings/layout.tsx
@@ -1,8 +1,0 @@
-export default function Layout({ modal, children }: { modal: React.ReactNode; children: React.ReactNode }) {
-  return (
-    <>
-      <div>{modal}</div>
-      <div>{children}</div>
-    </>
-  );
-}
